+++ CHANGES.md
@@ -22,7 +22,14 @@
 
 ### New Modules
 
+  * `XMonad.Hooks.WindowSwallowing`
+
+    A handleEventHook that implements window swallowing:
+    Hide parent windows like terminals when opening other programs (like image viewers) from within them,
+    restoring them once the child application closes.
+
   * `XMonad.Actions.TiledWindowDragging`
+
     An action that allows you to change the position of windows by dragging them around.
 
   * `XMonad.Layout.ResizableThreeColumns`